Implement logistic classifier

+++ y3p/teams/detect.py
@@ -1,14 +1,17 @@
 import cv2
 import os
+import sys
 import math
 import matplotlib.pyplot as plt
 import numpy as np
 
 from y3p import PROJECT_DIR
-from y3p.player.classification import TeamClassifier
 from y3p.space.camera import Camera
 
+FRAMES_TO_SKIP = 12
+
 def main(config, detector):
+  """Sample humans from videos and manually classify into team A, team B, and spectators"""
   classifier_config = config['team_classification']
   views = config['views']
   max_samples = classifier_config['samples']
@@ -33,17 +36,16 @@ def main(config, detector):
 
   stop = False
 
-  print('Press z for team A, x for team B, n for spectator...')
+  print('Press z for team A, x for team B, n for spectator, s to skip, b to go back, q to exit...')
 
-  teamA = 0
-  teamB = 0
+  team_a = 0
+  team_b = 0
   spectators = 0
+  frames_to_skip = 0
 
   while not stop:
     if samples >= max_samples:
       break
-
-    print('%d samples to do.' % (max_samples - samples))
 
     for i, capture in enumerate(captures):
       ret, frame = capture.read()
@@ -52,43 +54,64 @@ def main(config, detector):
         stop = True
         break
 
+      if frames_to_skip > 0:
+        frames_to_skip -= 1
+        continue
+
       detections = detector.forward(frame)
 
       for detection in detections:
-        samples += 1
         image = detection[4]
         mask = detection[5]
 
         img_path = os.path.join(PROJECT_DIR, out_dir, 'sample-%d.png' % samples)
-        mask_path = os.path.join(PROJECT_DIR, out_dir, 'sample-%d' % samples)
+        mask_path = os.path.join(PROJECT_DIR, out_dir, 'sample-%d.npy' % samples)
 
         cv2.imshow('Detection', image)
 
         key = cv2.waitKey(0) & 0xFF
 
         if key == ord('z'):
+          # team a
           labels.append(0)
-          teamA += 1
+          team_a += 1
         elif key == ord('x'):
+          # team b
           labels.append(1)
-          teamB += 1
+          team_b += 1
         elif key == ord('n'):
+          # spectator
           labels.append(2)
           spectators += 1
+        elif key == ord('s'):
+          continue
+        elif key == ord('b'):
+          print('Going back.')
+          samples -= 1
+          labels.pop()
+          continue
+        elif key == ord('q'):
+          print('Exiting.')
+          sys.exit(0)
         else:
-          labels.append(2)
-          spectators += 1
-          print('Key not recognised, assumed as spectator.')
+          print('Key not recognised, assumed as spectator. Skipping.')
+          continue
 
         cv2.destroyAllWindows()
         cv2.imwrite(img_path, image)
         np.save(mask_path, mask)
 
-  labels_path = os.path.join(PROJECT_DIR, out_dir, 'labels')
-  np.save(labels_path, labels)
+        samples += 1
+      
+      frames_to_skip = FRAMES_TO_SKIP
+      
+      print('%d samples to do.' % (max_samples - samples))
 
-  print('Saved images.')
-  print('Stats: %d in team A, %d in team B, %d spectators' % (teamA, teamB, spectators))
+  labels_path = os.path.join(PROJECT_DIR, out_dir, 'labels.npy')
+  np.save(labels_path, labels)
+  
+  print('%d samples in todal.' % samples)
+  print('Stats: %d in team A, %d in team B, %d spectators' % (team_a, team_b, spectators))
 
   # close video files
   for capture in captures:
